# PromQL: reduce garbage in range-query evaluation